Committing essential files.
In place of two separate files for the resume,
I combined them into one file. This was done
through the use of a "fontstyle" variable
in the documentclass declaration which gets
modified when the user invokes a make command.

In this way, users only need to modify *one* resume
file and can easily generate both pdfs using
both font styles. This method also generalizes
in the sense that many different style classes
can be created by the user, perhaps with different
fonts, and all that needs to be done to accomodate
compiling with those styles is to modify the makefile.