Commiting the makefile
The makefile is used to compile the resume.xtx
file into a pdf with some particular choice of
style.

The resume can be compiled by invoking the
"make foo" command, where foo is a fontstyle
given in the makefile. Currently we provide
support for "openfonts" and "macfonts".

Upon calling the make command, a pdf is output,
and any extraneous files are removed.